add linuxFxString error capability
